requirements.txt
Flask
python-dotenv
pymysql
python-dotenv
